% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction_df.R
\name{create_prediction_df}
\alias{create_prediction_df}
\title{Create prediction dataframe}
\usage{
create_prediction_df(
  model,
  plot_by = "doy",
  across_increment = 1,
  quant_trimming = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{model}{A GAM model object from mgcv, or a named list of fitted mgcv models}

\item{plot_by}{A single string specifying the variable to plot across. Defaults to "doy".}

\item{across_increment}{A numeric value specifying the increment for the plot_by variable. Defaults to 1.}

\item{quant_trimming}{A numeric value between 0 and 1 defining quantile-based trimming -- this provides a buffer between the outmost observations and the outmost plot predictions. Defaults to 0.01; increase if plot predictions at the edges are misleading.}

\item{verbose}{Provide context? Logical, defaults to TRUE}
}
\value{
tibble of variables relevant for the provided model, with all relevant combinations of
variable values.
}
\description{
From an MGCV model, create a dataframe (tibble) of values to predict across, including
all combinations of categorical predictors in the model, and incrementing across
the \code{plot_by} variable, and using median values of other numeric predictors.
Used by \code{predict_gam()} -- see there for more details
}
