% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_2md.R
\name{doy_2m}
\alias{doy_2m}
\title{Provide month from day of year. Uses 2019 as the model year.}
\usage{
doy_2m(i)
}
\arguments{
\item{i}{day of year}
}
\value{
character of form "Month"
}
\description{
Provide month from day of year. Uses 2019 as the model year.
}
\examples{
doy_2m(100)

}
