% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_2md.R
\name{doy_2md}
\alias{doy_2md}
\title{Provide month and day from day of year. Uses 2019 as the model year.}
\usage{
doy_2md(i)
}
\arguments{
\item{i}{day of year}
}
\value{
character of form "Month day"
}
\description{
Useful as labeling function in scale_x_continuous or scale_y_continuous
}
\examples{
doy_2md(100)
}
