% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gam.R
\name{predict_gam}
\alias{predict_gam}
\title{Predict GAM across reasonable values}
\usage{
predict_gam(
  model,
  plot_by = "doy",
  across_increment = 1,
  quant_trimming = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{model}{A GAM model object from mgcv, or a named list of fitted mgcv models}

\item{plot_by}{A single string specifying the variable to plot across. Defaults to "doy".}

\item{across_increment}{A numeric value specifying the increment for the plot_by variable. Defaults to 1.}

\item{quant_trimming}{A numeric value between 0 and 1 defining quantile-based trimming -- this provides a buffer between the outmost observations and the outmost plot predictions. Defaults to 0.01; increase if plot predictions at the edges are misleading.}

\item{verbose}{Provide context? Logical, defaults to TRUE}
}
\value{
A data frame with predictions, confidence intervals, and predictor values.
}
\description{
Predicts gam output. Primarily an internal tool, used in \code{plot_seasonal_gam_panels.R}.
}
