% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate fish count data}
\usage{
simulate_data(
  n.per.year = 60,
  years = 2021:2023,
  doy.limits = c(200, 300),
  middle.doy = 250,
  samples_per_day_limits = c(4, 10)
)
}
\arguments{
\item{n.per.year}{A single numeric value defining the number of samples per year. Defaults to 60.}

\item{years}{A numeric vector defining a "years" factor. Defaults to 2021:2023}

\item{doy.limits}{A numeric vector defining minimum and maximum day of year to simulate sampling.
Length 2. Optional.}

\item{middle.doy}{A numeric defining the day of year of the general peak of activity. Defaults to 250.}

\item{samples_per_day_limits}{A numeric vector of length 2. Optional.}
}
\value{
A data frame with columns for day of year, year, mark stage, year factor,
and simulated fish counts.
}
\description{
Simple function to generate reasonable-shaped creel sample data. Useful for testing / demonstrating plotting functions. This can (and typically should) be used without providing any arguments; several
aspects of the simulated data (e.g. # of years simulated, number of sampeles) can be modified with the optional arguments.
}
