% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_gam}
\alias{simulate_gam}
\title{Simulate mgcv gam model for demonstration purposes}
\usage{
simulate_gam()
}
\value{
fitted mgcv model
}
\description{
Uses \code{simulate_data()}, and fits a negative binomial gam model for response \code{fish_count}  using
predictor variables \code{mark_stage} ("Jack", "AD Adult", "UM Adult"), \code{yearfac} (factor version of year), and a smooth across \code{doy} (numeric for day of year). Output is suitable model for
\code{plot_seasonal_gam_panels()}.
}
\examples{
\dontrun{
model = simulate_gam()
model_panels = plot_seasonal_gam_panels(model)
wrap_panels(model_panels)
}
}
