% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{validate_number}
\alias{validate_number}
\title{Test argument to see if it's a numeric within one more constraints.}
\usage{
validate_number(
  x,
  limits_inclusive = NULL,
  limits_exclusive = NULL,
  ...,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{item to test}

\item{limits_inclusive}{Defines inclusive constraints. If provided, numeric vector of length 2, with first value being the lower limit and second value being the upper limit. Use NA to identify a limit that should be ignored. E.g., if a value should be no less than 0 but has no upper limit, use \code{limits_inclusive = c(0, NA)}.}

\item{limits_exclusive}{As limits_inclusive, but for exclusive limits (e.g., a value must be greater than 0).}

\item{...}{additional arguments}

\item{arg}{For identifying argument from the calling function}

\item{call}{For identifying the calling function}
}
\value{
None.
}
\description{
Can identify lower and/or upper limits, either inclusive (\code{limits_inclusive}) or exclusive
(\code{limits_exclusive}).
}
