% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizSeasonalGams-package.R
\docType{package}
\name{VizSeasonalGams-package}
\alias{VizSeasonalGams}
\alias{VizSeasonalGams-package}
\title{VizSeasonalGams: Visualize Fitted MGCV Models With Seasonal Smooths}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/VizSeasonalGams}
  \item \url{https://cbedwards-dfw.github.io/VizSeasonalGams/}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/VizSeasonalGams/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
