% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_panels.R
\name{wrap_panels}
\alias{wrap_panels}
\title{Wrap panels from seasonal GAM plots}
\usage{
wrap_panels(panel_df, ncol = NULL, nrow = NULL, title_size = 8)
}
\arguments{
\item{panel_df}{A dataframe from plot_seasonal_gam_panels.}

\item{ncol}{Optional. Force a number of columns. Numeric, defaults to NULL.}

\item{nrow}{Optional. Force a number of rows. Numeric, defaults to NULL.}

\item{title_size}{size of panel title text. Numeric, defaults to 8.}
}
\value{
A wrapped plot object with row and column labels_
}
\description{
Analog to facet_wrap() for the seasonal gam panels. If there are two predictors
of interest, consider \code{grid_panels()} instead.
}
