% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_panels.R
\name{grid_panels}
\alias{grid_panels}
\title{Title}
\usage{
grid_panels(
  panels_df,
  column_var,
  row_var,
  panel_titles = FALSE,
  col_title_ratio = 30,
  row_title_ratio = 30,
  font_size = 18,
  new_y_label = NULL,
  new_x_label = NULL,
  title_size = 8
)
}
\arguments{
\item{panels_df}{A dataframe from plot_seasonal_gam_panels.}

\item{column_var}{1 dimensional character vector of variable defining the columns of the grid. Must be factor variable present in \code{panels_df}.}

\item{row_var}{as \code{column_var}, but defines the rows of the grid}

\item{panel_titles}{Include all panel information in panel title even if redundant with rows/panels? Useful for debugging or screenshotting sections. Logical, defaults to false}

\item{col_title_ratio}{How much bigger should content be than the column labels? Numeric, defaults to 30.}

\item{row_title_ratio}{as \code{col_title_ratio}, but ratio of content to row labels.}

\item{font_size}{Size of row and column labels. Numeric, defaults to 18.}

\item{new_y_label}{Optional, replaces Y axis label of primary plots in panels (not data coverage histograms). character vector, defaults to NULL.}

\item{new_x_label}{Optional, replaces X axis label of panels. Character vector, defaults to NULL.}

\item{title_size}{size of panel title text. Numeric, defaults to 8.}
}
\value{
patchwork object
}
\description{
Title
}
