% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_gam_panels.R
\name{plot_seasonal_gam_panels}
\alias{plot_seasonal_gam_panels}
\title{Create tibble of predictors and prediction plots}
\usage{
plot_seasonal_gam_panels(
  model,
  color_by = NULL,
  plot_across = "doy",
  across_increment = 1,
  quant_trimming = 0.01,
  include_cis = TRUE,
  plot_coverage = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{A GAM model object from mgcv}

\item{color_by}{Variable to color predictions by. Defaults to NULL, in which case curves are black and each panel has only one set of predictions. When used, color_by necessarily reduces dimensionalality of output by 1.}

\item{plot_across}{A single string specifying the variable to plot across. Defaults to "doy".}

\item{across_increment}{A numeric value specifying the increment for the plot_across variable. Defaults to 1.}

\item{quant_trimming}{A numeric value between 0 and 1 defining quantile-based trimming -- this provides a buffer between the outmost observations and the outmost plot predictions. Defaults to 0.01; increase if plot predictions at the edges are misleading.}

\item{include_cis}{Should we plot the confidence envelope? Logical, defaults to TRUE}

\item{plot_coverage}{Add histogram at bottom of each panel with data coverage? Logical, defaults to TRUE.}

\item{verbose}{Provide context? Logical, defaults to TRUE}
}
\value{
Tibble, with a column for each factor predictor in \code{model}. \verb{$plot} column contains the plot panels.
}
\description{
Creates a tibble with all represented combinations of factor variables in the data,
and includes a model of model predictions across a key continuous variable (\code{plot_across}, defaults to "doy")
while setting other continuous variables to their median. Because these plots are within a tibble framework, it is easy to filter to desired combinations of
variables and plot those. (e.g., model might fit 20 years for 13 spatial units, tibble
can rapidly be filtered to just the years of a desired spatial unit, or all spatial units
of a desired year).
}
